13MAI2022

+++ Blog/public/delete_article.php
@@ -1,5 +1,12 @@
 <?php
+session_start();
 include '../lib/functions.php';
+
+if(!hasRole(ROLE_ADMIN)) {
+    http_response_code(403);
+    echo 'Accès interdit';
+    exit;
+}
 
 $id = $_GET['id'];
 deleteArticle($id);
